--- tests/TestHelper.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Luma\RoutingComponentTests;
-
-class TestHelper
-{
-
-}
