+++ tests/Unit/RouterTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Luma\RoutingComponentTests;
+namespace Luma\Tests;
 
 use Luma\HttpComponent\Request;
 use Luma\HttpComponent\URI;
 use Luma\RoutingComponent\Router;
+use Luma\Tests\TestHelper;
 use PHPUnit\Framework\MockObject\Exception as MockObjectException;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
