+++ tests/Unit/TestHelper.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Luma\Tests;
+
+class TestHelper
+{
+
+}
